adding species
