+++ src/aind_data_schema_models/species.py
@@ -2,76 +2,25 @@
 
 from typing import Literal, Union
 
-from pydantic import ConfigDict, Field
+from pydantic import ConfigDict, Field, BaseModel
 from typing_extensions import Annotated
+import pandas as pd
+from aind_data_schema_models.utils import create_literal_class
+from aind_data_schema_models.registry import Registry, Registries, map_registry
 
-from aind_data_schema_models.pid_names import PIDName
-from aind_data_schema_models.registry import NationalCenterForBiotechnologyInformation, Registry
-
-
-class _Species(PIDName):
-    """Base model config"""
-
+class SpeciesModel(BaseModel):
     model_config = ConfigDict(frozen=True)
 
+    name: str
+    registry: Registry
+    registry_identifier: str
+    
 
-class CallithrixJacchus(_Species):
-    """Callithrix Jacchus"""
-
-    name: Literal["Callithrix jacchus"] = "Callithrix jacchus"
-    registry: Annotated[
-        Union[NationalCenterForBiotechnologyInformation], Field(default=Registry.NCBI, discriminator="name")
-    ]
-    registry_identifier: Literal["NCBI:txid9483"] = "NCBI:txid9483"
-
-
-class HomoSapiens(_Species):
-    """Homo Sapiens"""
-
-    name: Literal["Homo sapiens"] = "Homo sapiens"
-    registry: Annotated[
-        Union[NationalCenterForBiotechnologyInformation], Field(default=Registry.NCBI, discriminator="name")
-    ]
-    registry_identifier: Literal["NCBI:txid9606"] = "NCBI:txid9606"
-
-
-class MacacaMulatta(_Species):
-    """Macaca Mulatta"""
-
-    name: Literal["Macaca mulatta"] = "Macaca mulatta"
-    registry: Annotated[
-        Union[NationalCenterForBiotechnologyInformation], Field(default=Registry.NCBI, discriminator="name")
-    ]
-    registry_identifier: Literal["NCBI:txid9544"] = "NCBI:txid9544"
-
-
-class MusMusculus(_Species):
-    """Mus Musculus"""
-
-    name: Literal["Mus musculus"] = "Mus musculus"
-    registry: Annotated[
-        Union[NationalCenterForBiotechnologyInformation], Field(default=Registry.NCBI, discriminator="name")
-    ]
-    registry_identifier: Literal["NCBI:txid10090"] = "NCBI:txid10090"
-
-
-class RattusNorvegicus(_Species):
-    """Rattus Norvegicus"""
-
-    name: Literal["Rattus norvegicus"] = "Rattus norvegicus"
-    registry: Annotated[
-        Union[NationalCenterForBiotechnologyInformation], Field(default=Registry.NCBI, discriminator="name")
-    ]
-    registry_identifier: Literal["NCBI:txid10116"] = "NCBI:txid10116"
-
-
-class Species:
-    """Species classes"""
-
-    CALLITHRIX_JACCHUS = CallithrixJacchus()
-    HOMO_SAPIENS = HomoSapiens()
-    MACACA_MULATTA = MacacaMulatta()
-    MUS_MUSCULUS = MusMusculus()
-    RATTUS_NOVEGICUS = RattusNorvegicus()
-    _ALL = tuple(_Species.__subclasses__())
-    ONE_OF = Annotated[Union[_ALL], Field(discriminator="name")]
+Species = create_literal_class(
+    objects=pd.read_csv('models/species.csv', keep_default_na=False).to_dict(orient='records'), 
+    class_name='Species', 
+    base_model=SpeciesModel, 
+    discriminator='name',
+    field_handlers={'registry_abbreviation': map_registry},
+    class_module=__name__
+)
